+++ matchmaker-cli/src/paths.rs
@@ -1,24 +1,24 @@
 use std::path::PathBuf;
 
-use cli_boilerplate_automation::expr_as_path_fn;
+use cba::expr_as_path_fn;
 
-use crate::clap::BINARY_FULL;
+use crate::clap::LIBRARY_FULL;
 
 fn config_dir_impl() -> Option<PathBuf> {
     if let Some(home) = dirs::home_dir() {
-        let config = home.join(".config").join(BINARY_FULL);
+        let config = home.join(".config").join(LIBRARY_FULL);
         if config.exists() {
             return Some(config);
         }
     };
 
-    dirs::config_dir().map(|x| x.join(BINARY_FULL))
+    dirs::config_dir().map(|x| x.join(LIBRARY_FULL))
 }
 
 pub fn state_dir_impl() -> Option<PathBuf> {
     dirs::state_dir()
         .or_else(|| dirs::home_dir().map(|home| home.join(".local").join("state")))
-        .map(|x| x.join(BINARY_FULL))
+        .map(|x| x.join(LIBRARY_FULL))
 }
 
 expr_as_path_fn!(state_dir, state_dir_impl().unwrap_or_default());