+++ matchmaker-cli/src/start.rs
@@ -2,7 +2,6 @@ use std::{
     io::Read,
     path::Path,
     process::{Command, exit},
-    sync::Arc,
 };
 
 use crate::{
@@ -12,7 +11,7 @@ use crate::{
     paths::last_key_path,
 };
 use crate::{config::Config, paths::default_config_path};
-use cli_boilerplate_automation::{
+use cba::{
     bait::{OptionExt, ResultExt},
     bo::{
         MapReaderError, load_type_or_default, map_chunks, map_reader_lines, read_to_chunks,
@@ -21,7 +20,7 @@ use cli_boilerplate_automation::{
     bog::BogOkExt,
     prints,
 };
-use cli_boilerplate_automation::{bo::load_type, broc::CommandExt};
+use cba::{bo::load_type, broc::CommandExt};
 use log::debug;
 use matchmaker::{
     ConfigInjector, MatchError, Matchmaker, OddEnds, PickOptions, SSS,
@@ -32,9 +31,11 @@ use matchmaker::{
     message::Interrupt,
     nucleo::{
         ColumnIndexable,
-        injector::{AnsiInjector, IndexedInjector, Injector, SegmentedInjector},
+        injector::{AnsiInjector, Either, IndexedInjector, Injector, SegmentedInjector},
     },
     preview::AppendOnly,
+    render::MMState,
+    use_formatter,
 };
 use matchmaker_partial::Apply;
 
@@ -139,7 +140,7 @@ pub fn map_reader<E: SSS + std::fmt::Display>(
     })
 }
 
-use cli_boilerplate_automation::wbog;
+use cba::wbog;
 use matchmaker::{Actions, binds::Trigger};
 use std::collections::HashMap;
 pub fn resolve_aliases(
@@ -212,40 +213,56 @@ pub async fn start(config: Config, no_read: bool) -> Result<(), MatchError> {
         mut mm,
         injector,
         OddEnds {
-            formatter,
             splitter,
             hidden_columns,
         },
     ) = Matchmaker::new_from_config(render, tui, worker, exit, preprocess);
     // make previewer
     let help_str = display_binds(&event_loop.binds, Some(&previewer.help_colors));
-    let previewer = make_previewer(&mut mm, previewer, formatter.clone(), help_str);
+    let cli_formatter = Either::Right(
+        crate::formatter::format_cli
+            as for<'a, 'b, 'c> fn(
+                &'a MMState<'b, 'c, matchmaker::ConfigMMItem, matchmaker::ConfigMMInnerItem>,
+                &'a str,
+                Option<&dyn Fn(String)>,
+            ) -> String,
+    );
+    let previewer = make_previewer(&mut mm, previewer, cli_formatter.clone(), help_str);
 
     // ---------------------- register handlers ---------------------------
     // print handler (no quoting)
-    let print_formatter = Arc::new(mm.worker.default_format_fn::<false>(|item| item.to_cow()));
     mm.register_print_handler(
         print_handle.clone(),
         output_separator.clone(),
-        print_formatter.clone(),
+        cli_formatter.clone(),
     );
 
     // execute handlers
-    mm.register_execute_handler(formatter.clone());
-    mm.register_become_handler(formatter.clone());
+    mm.register_execute_handler(cli_formatter.clone());
+    mm.register_become_handler(cli_formatter.clone());
 
     // reload handler
-    let reload_formatter = formatter.clone();
+    let reload_formatter = cli_formatter.clone();
+    let default_reload = (!command.is_empty() && atty::is(atty::Stream::Stdin) || no_read)
+        .then_some(command.clone())
+        .unwrap_or_default();
+
     mm.register_interrupt_handler(Interrupt::Reload, move |state| {
         let injector = state.injector();
         let injector = IndexedInjector::new_globally_indexed(injector);
         let injector = SegmentedInjector::new(injector, splitter.clone());
         let injector = AnsiInjector::new(injector, preprocess);
 
-        if let Some(t) = state.current_raw() {
-            let cmd = reload_formatter(t, state.payload());
+        let cmd = if !state.payload().is_empty() {
+            &use_formatter(&reload_formatter, state, state.payload(), None)
+        } else {
+            &default_reload
+        };
+
+        if !cmd.is_empty() {
             let vars = state.make_env_vars();
             debug!("Reloading: {cmd}");
+            state.picker_ui.selector.clear();
             if let Some(stdout) = Command::from_script(&cmd).envs(vars).spawn_piped()._elog() {
                 map_reader(
                     stdout,
@@ -262,6 +279,9 @@ pub async fn start(config: Config, no_read: bool) -> Result<(), MatchError> {
     let mut action_context = ActionContext {
         bind_tx: event_loop.bind_controller(),
         additional_commands: (additional_commands, 0),
+        output_template,
+        print_handle: print_handle.clone(),
+        output_separator: output_separator.clone(),
     };
 
     let mut options = PickOptions::new()
@@ -269,7 +289,11 @@ pub async fn start(config: Config, no_read: bool) -> Result<(), MatchError> {
         .matcher(matcher.0)
         .previewer(previewer)
         .hidden_columns(hidden_columns)
-        .ext_handler(move |x, y| action_handler(x, y, &mut action_context));
+        .ext_handler(move |x, y| action_handler(x, y, &mut action_context))
+        .ext_aliaser(|a, _state| match a {
+            matchmaker::Action::Accept => matchmaker::acs![MMAction::Accept],
+            _ => matchmaker::acs![a],
+        });
 
     let render_tx = options.render_tx();
 
@@ -301,26 +325,13 @@ pub async fn start(config: Config, no_read: bool) -> Result<(), MatchError> {
         handle.await._wbog(); // warn the mapreader error (?)
     }
 
-    mm.pick(options).await.map(|v| {
-        print_handle.map_to_vec(|s| print!("{}{}", s, output_separator));
-
-        for item in v {
-            let s = if let Some(s) = &output_template {
-                print_formatter(
-                    &matchmaker::nucleo::Indexed {
-                        index: 0,
-                        inner: item,
-                    },
-                    s,
-                )
-                .into()
-            } else {
-                item.to_cow()
-            };
-
-            print!("{}{}", s, output_separator)
+    match mm.pick(options).await {
+        Ok(_) | Err(MatchError::NoMatch) => {
+            print_handle.map_to_vec(|s| print!("{}{}", s, output_separator));
+            Ok(())
         }
-    })
+        Err(e) => Err(e),
+    }
 }
 
 use matchmaker::nucleo::{Line, Span};