+++ matchmaker-lib/src/action.rs
@@ -104,6 +104,8 @@ pub enum Action<A: ActionExt = NullActionExt> {
     // Programmable
     /// Execute command and continue
     Execute(String),
+    /// Execute command without leaving the UI
+    ExecuteSilent(String),
     /// Exit and become
     Become(String),
     /// Reload matcher/worker
@@ -346,11 +348,11 @@ enum_from_str_display!(
     ForwardChar,BackwardChar, ForwardWord, BackwardWord, DeleteChar, DeleteWord, DeleteLineStart, DeleteLineEnd, Cancel, Redraw, NextColumn, PrevColumn;
 
     tuples:
-    Execute, Become, Reload, Preview,
+    Execute, ExecuteSilent, Become, Preview,
     SetQuery, Pos, QueryPos, SwitchColumn;
 
     defaults:
-    (Up, 1), (Down, 1), (PreviewUp, 1), (PreviewDown, 1), (Quit, 1), (Overlay, 0), (Print, String::new()), (Help, String::new()), (PreviewScroll, 1), (PreviewHScroll, 1), (HScroll, 0), (VScroll, 0);
+    (Up, 1), (Down, 1), (PreviewUp, 1), (PreviewDown, 1), (Quit, 1), (Overlay, 0), (Print, String::new()), (Help, String::new()), (Reload, String::new()), (PreviewScroll, 1), (PreviewHScroll, 1), (HScroll, 0), (VScroll, 0);
 
     options:
     SwitchPreview, SetPreview, ToggleColumn, ShowColumn