+++ matchmaker-lib/src/render/mod.rs
@@ -434,6 +434,9 @@ pub(crate) async fn render_loop<'a, W: Write, T: SSS, S: Selection, A: ActionExt
                         Action::Execute(payload) => {
                             state.set_interrupt(Interrupt::Execute, payload);
                         }
+                        Action::ExecuteSilent(payload) => {
+                            state.set_interrupt(Interrupt::ExecuteSilent, payload);
+                        }
                         Action::Become(payload) => {
                             state.set_interrupt(Interrupt::Become, payload);
                         }
@@ -511,7 +514,9 @@ pub(crate) async fn render_loop<'a, W: Write, T: SSS, S: Selection, A: ActionExt
 
                         Action::ShowColumn(col_name) => {
                             if let Some(name) = col_name {
-                                if let Some(idx) = worker.columns.iter().position(|c| *c.name == name) {
+                                if let Some(idx) =
+                                    worker.columns.iter().position(|c| *c.name == name)
+                                {
                                     if idx < results.hidden_columns.len() {
                                         results.hidden_columns[idx] = false;
                                     }