+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.0.12] - 2026-03-09
+
+### 🚀 Features
+
+- Cleaner help display
+- Column styles
+- Finalize templating
+- `matcher.start.default_column` and `matcher.start.additional_commands`
+- ExecuteSilent action
+- various bugfixes and documentation
+
+### Performance
+- Streamline AppendOnly (preview synchronization) using arc-swap
+
 ## [0.0.10] - 2026-03-07
 
 ### 💼 Other
@@ -41,11 +55,13 @@
 ## [0.0.7] - 2026-02-22
 
 ### 🚀 Features
+
 - matchmaker-partial: support recursive set in collections
 - matchmaker-cli: support direct override of preview command (alias: px)
 - matchmaker-cli: new aliases: see options.md
 
 ### 🚜 Refactor
+
 - Move start and exit configs out from under MatcherConfig to top level
 
 ## [0.0.6] - 2026-02-22
@@ -57,15 +73,18 @@
 ### 🚜 Refactor
 
 - Lints
+
 ## [0.0.4] - 2026-02-19
+
 - Bugfix and documentation updates
 - Align version cli and library versions
 
 ## [0.0.2] - 2026-02-18
+
 - Various bugfixes and improvements
 - New configuration options:
-	- PreviewScrollSetting
-	- print_template
+  - PreviewScrollSetting
+  - print_template
 
 ## [0.0.1] - 2026-02-16